backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
sentence-transformers==2.2.2
torch==2.0.1+cpu
torchvision==0.15.2+cpu
torchaudio==2.0.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
faiss-cpu
python-dotenv
httpx
pytest
python-multipart
passlib[bcrypt]
PyJWT
pandas
reportlab
jinja2
amadeus
